Refactor

+++ dataset.py
@@ -101,10 +101,10 @@ def gen(coco):
 
 
 def rescale_image(image, scale):
-    size = tf.shape(image)[:2]
+    size = tf.to_float(tf.shape(image)[:2])
     shorter = tf.argmin(size)
     ratio = scale / size[shorter]
-    new_size = tf.to_int32(tf.round(tf.to_float(size) * ratio))
+    new_size = tf.to_int32(tf.round(size * ratio))
 
     return tf.image.resize_images(image, new_size, method=tf.image.ResizeMethod.BILINEAR)
